feat: simplifying exam verify utilities usage
