+++ client/src/css/Exam.css
@@ -106,8 +106,8 @@
 
 .verify-box {
     width: 122px;
-    margin-top: 40px;
-    margin-bottom: 40px;
+    margin-top: 20px;
+    margin-bottom: 20px;
     padding: 20px;
     text-align: center;
     background-color: #fff;
